Add selection variables and function playRound()
Starting the exercise from scratch and figuring out the algorithm
before actually starting to code really really helped in understanding
what the code meant to the machine and I also could figure out
the error I was doing in the previous version of the exercise:
I was calling out the function (that had parameters), without
putting any arguments into the function. In that case the arguments
were UNDEFINED and the strict equality between the two parameters always
returned true.